requirement.txt
tensorflow==1.15.5
tflearn==0.5.0
numpy>=1.20
pillow>=10.0.0
matplotlib>=3.7
opencv-python
scipy
pandas